connected to endpoints

+++ src/components/RegisterModal.jsx
@@ -1,21 +1,21 @@
 import { Modal, Input, Button } from "@mantine/core"
-import { register } from "../api/register.js"
+import { register } from "../api/apicontrollers.js"
 import React, { useState } from 'react';
 import "../styles/modal.scss"
 
 export default function RegisterModal({opened, setOpened}) {
 
   const [username, setUsername] = useState("")
+  const [alert, setAlert] = useState("")
 
-  // Set username Local Storage
-  function setUsernameLS(username) {
-	localStorage.setItem("username",username)
-	setOpened(false)
-  }
-
-  async function handleRegister(username) {
+  async function handleRegister() {
 	let checkUsername = await register(username)
-	console.log("Checkusername ",checkUsername)
+	if(checkUsername.data.message === "Username already taken"){
+	  setAlert("Username already taken")
+	} else {
+	  localStorage.setItem("username", username)
+	  setOpened(false)
+	}
   }
 
   return (
@@ -28,6 +28,7 @@ export default function RegisterModal({opened, setOpened}) {
 	  >
 		<div className="modal-main-container">
 		  <h2>Register</h2>
+		  <h3>{alert}</h3>
 		  <Input
 			placeholder="Username"
 			onChange={(e)=>setUsername(e.target.value)}
@@ -35,7 +36,6 @@ export default function RegisterModal({opened, setOpened}) {
 			  (e)=>{
 				if(e.keyCode === 13){
 					handleRegister(username)
-					//setUsernameLS(username)
 				}
 			  }
 			}
@@ -44,7 +44,6 @@ export default function RegisterModal({opened, setOpened}) {
 			  className="button"
 			  onClick={()=>{
 				handleRegister(username)
-				//setUsernameLS(username)
 			  }}>
 			Save
 		  </Button>
